% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseControlValue.R
\name{chooseControlValue}
\alias{chooseControlValue}
\title{Choose control variable value}
\usage{
chooseControlValue(varName, defaultVal, inputControl, haveChain,
  prevControl = NA)
}
\arguments{
\item{varName}{The variable name}

\item{defaultVal}{The default value}

\item{inputControl}{The user specified control list}

\item{haveChain}{A boolean indicating whether the values from a previous run are available}

\item{prevControl}{The control list from a previous run}
}
\value{
The preferred value for the variable
}
\description{
This "helper" function chooses the value for a control variable that can either be (in order of increasing preference) the default, the value fro a previous run, and a directly input value.
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
