% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam_methods.R
\name{print.sam}
\alias{print.sam}
\title{Print a \code{sam} object}
\usage{
\method{print}{sam}(x)
}
\arguments{
\item{x}{The \code{sam} object to print}
}
\value{
Silently returns the input \code{chain} object
}
\description{
Place holder
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
