% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samMCMC.R
\name{samMCMC}
\alias{samMCMC}
\title{Adaptive Metropolis sampling}
\usage{
samMCMC(sampFunc, init, temp = NA, ..., control = list())
}
\arguments{
\item{init}{The starting point for sampling, X_0, or the output from a previous call 
to samMCMC}

\item{temp}{The temperature for sampling}

\item{...}{Further arguments to be passed to costFunc}

\item{control}{A list of control parameters. See Details.}

\item{costFunc}{The cost function (often a negative log-likelihood)}
}
\value{
An object of class \code{chain} that is a list containing the samples along 
with summary information
}
\description{
This function samples using the Metropolis algorithm described by
Harrio, Saksman, and Tamminen (2001). \code{samMCMC} works in one of two ways:

(1) [default] The function to be sampled is directly specified
(2)           The negative log density of the function to be sampled is specified,
              along with a temperature
}
\details{
The first input to the cost functionn, sampFunc, must be a function taking
 the variable to be sampled as its first input. Additional inputs to sampmFunc can be
 given as named variable inputs.

The \code{control} argument is a list that can supply any of the following (otherwise 
defaults, in brackets, are be used):
\itemize{
  \item{\code{numSamp}} Number of samples after initialization [1000]
  \item{\code{t0}} Number of samples using initial proporsal ditribution [100]
  \item{\code{verbose}} Whether to print out information as the run proceeds [F]
  \item{\code{fileName}} Filename for saving
  \item{\code{savePeriod}} Period (of samples) for saving [1000]
}
}
\references{
Harrio, Saksman, and Tamminen (2001), An adaptive Metropolis algorithm, 
in Bernoulli, Volume 7, Number 2, pages 223 through 242.
}
\author{
Michael Holton Price <MichaelHoltonPrice@gmail.com>
}
